backend/requirements.txt
asgiref==3.11.1
Django==5.2.12
wheel==0.46.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
openpyxl==3.1.2
gunicorn
twilio
git-filter-repo==2.47.0
django-environ
setuptools>=80.0.0
